mqtt-py/mqtt/requirements.txt
paho-mqtt
requests
discordwebhook
python-dotenv